pyproject.toml
[project]
name = "homesecurity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "eventlet>=0.40.3",
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "paho-mqtt>=2.1.0",
    "requests>=2.32.5",
]
